refactor: refactored the sortable component
