+++ src/lib/types/formSelectItemstype.ts
@@ -10,3 +10,9 @@ export interface FormSelectItem {
 
 export type IconMapType = Record<iconType, LucideIcon>;
 export type QueryType = "input" | "dropdown";
+
+export interface AddedInput {
+  id: string;
+  type: QueryType;
+  label: string;
+}
